requirements.txt
mfrc522==0.0.7
RPi.GPIO==0.7.0
